Tooltip
* Removed percentages from tooltips
* display counts if no targets otherwise [count]/[target]
* Moved functions that were shared by the graph chart because they will
  be reconfigured and out of scope for this
* Started counts as 0 in the chance that the site is new and there is no
  data
